requirements.txt
python-telegram-bot>=20.0
httpx>=0.23.1,<0.24.0
asyncpg>=0.27.0
streamlit>=1.10.0
plotly>=5.0.0
pandas>=1.3.0
psycopg2-binary>=2.9.0
